Massive split of Bits into BitsRef and new Bits type
