+++ Cargo.toml
@@ -38,6 +38,7 @@ rand_xorshift = "0.1"
 criterion = "0.2"
 bincode = "1"
 proptest = "0.8"
+quickcheck = "0.7"
 
 [[bench]]
 name = "operations"
